remove improvement of the mapping information
