+++ spring-boot-project/spring-boot-actuator/src/main/java/org/springframework/boot/actuate/web/mappings/servlet/DispatcherServletMappingDescription.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2025 the original author or authors.
+ * Copyright 2012-2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 
 package org.springframework.boot.actuate.web.mappings.servlet;
 
-import java.util.List;
-
 import org.springframework.web.servlet.DispatcherServlet;
 
 /**
@@ -32,22 +30,12 @@ public class DispatcherServletMappingDescription {
 
 	private final String predicate;
 
-	private final List<DispatcherServletMappingDescription> children;
-
 	private final DispatcherServletMappingDetails details;
-
-	public DispatcherServletMappingDescription(String predicate, String handler, DispatcherServletMappingDetails details, List<DispatcherServletMappingDescription> children) {
-		this.handler = handler;
-		this.predicate = predicate;
-		this.details = details;
-		this.children = children;
-	}
 
 	DispatcherServletMappingDescription(String predicate, String handler, DispatcherServletMappingDetails details) {
 		this.handler = handler;
 		this.predicate = predicate;
 		this.details = details;
-		this.children = null;
 	}
 
 	public String getHandler() {
@@ -56,10 +44,6 @@ public class DispatcherServletMappingDescription {
 
 	public String getPredicate() {
 		return this.predicate;
-	}
-
-	public List<DispatcherServletMappingDescription> getChildren() {
-		return this.children;
 	}
 
 	public DispatcherServletMappingDetails getDetails() {
